rpiserver/pyproject.toml
[tool.poetry]
name = "rpiserver"
version = "0.1.0"
description = ""
authors = ["Paul Lecomte <paullecomte07@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0.53.2"
uvicorn="^0.11.3"
"RPi.GPIO" = ">=0.7.0"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
